% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cmdstan.R
\name{check_cmdstan}
\alias{check_cmdstan}
\title{Check CmdStan and its toolchain}
\usage{
check_cmdstan()
}
\value{
Invisibly returns a list with CmdStan version and toolchain check results.
}
\description{
Checks whether CmdStan is installed and the toolchain is working.
}
