% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{.get_posterior}
\alias{.get_posterior}
\title{Extract posterior samples}
\usage{
.get_posterior(fit)
}
\arguments{
\item{fit}{Named list of CmdStanMCMC objects}
}
\value{
Named list of tibbles containing posterior samples for each model
}
\description{
Extracts posterior samples from fitted models as a named list of tibbles
}
