% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{est_max}
\alias{est_max}
\title{Estimate maximum size}
\usage{
est_max(fit, ci = 0.8, k = 20)
}
\arguments{
\item{fit}{Fitted model object}

\item{ci}{Credible interval width (default: 0.8)}

\item{k}{The k-sample LMAX. By default this is set to 20, for ease of comparison between studies and species. Note that this is not the number of sample maxima used to fit the model. (default: 20)}
}
\value{
Tibble of maximum length, lower and upper confidence intervals
}
\description{
Estimates maximum , with uncertainty, size given k-samples
}
