% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{est_max_efs}
\alias{est_max_efs}
\title{Quantile Estimates from Truncated Normal-Based Posterior}
\usage{
est_max_efs(fitted_mod, q = 0.95)
}
\arguments{
\item{q}{Numeric scalar giving the quantile probability to evaluate
(default = 0.95).}

\item{posterior}{A fitted model object, created using the fit_mod(model_type = "efs") function}
}
\value{
A tibble with three summary statistics of the posterior quantile
distribution:
\item{fit}{Posterior median of the quantile}
\item{lwr}{10th percentile of the quantile distribution}
\item{upr}{90th percentile of the quantile distribution}
}
\description{
Computes posterior predictive quantiles for maximum values using a
truncated-normal-based model of fish size distributions.
}
\details{
This function requires an \code{inverse_cdf_x()} function to be
defined, which computes quantiles for the truncated normal extreme value
model.
}
\examples{
\dontrun{
est_max_efs(efs_fitted_model, q = 0.95)
}

}
