% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{est_max_evt}
\alias{est_max_evt}
\title{Quantile Estimates from a GEV Posterior}
\usage{
est_max_evt(evt_fitted_mod, q = 0.95)
}
\arguments{
\item{q}{Numeric scalar giving the quantile probability to evaluate
(default = 0.95).}

\item{evt_fitted_model}{A fitted model object, created using the fit_mod(model_type = "evt") function}
}
\value{
A tibble with three summary statistics of the posterior quantile
distribution:
\item{fit}{Posterior median of the quantile}
\item{lwr}{10th percentile of the quantile distribution}
\item{upr}{90th percentile of the quantile distribution}
}
\description{
Computes posterior predictive quantiles for extreme values using the
Generalized Extreme Value (GEV) distribution parameters sampled from a
posterior distribution.
}
\examples{
\dontrun{
est_max_evt(evt_fitted_model, q = 0.99)
}

}
