% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishEVA-package.R
\docType{package}
\name{fishEVA-package}
\alias{fishEVA-package}
\alias{fishEVA}
\title{fishEVA: Extreme Value Analysis for Marine Biology}
\description{
This package provides tools for extreme value analysis specifically
designed for marine biological data. It implements both traditional
Extreme Value Theory (EVT) and Extreme Fish Size (EFS) models using
Bayesian methods via Stan.
}
\section{Main functions}{

\describe{
\item{\code{\link{fit_evt_model}}}{Fit EVT or EFS models to data}
\item{\code{\link{get_posterior}}}{Extract posterior samples}
\item{\code{\link{est_max}}}{Estimate maximum size for given return period}
\item{\code{\link{get_pdf}}}{Compute probability density function}
\item{\code{\link{plot_model_comparison}}}{Create model comparison plots}
}
}

\author{
\strong{Maintainer}: Freddie Heather \email{freddieheather@gmail.com}

}
