% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishmax-package.R
\docType{package}
\name{fishmax-package}
\alias{fishmax-package}
\alias{fishmax}
\title{fishmax: Extreme Value Analysis for Estimating the maximum body length of fishes}
\description{
This package provides tools for extreme value analysis specifically
designed for fish body max length. It implements both traditional
Extreme Value Theory (EVT) and Extreme Fish Size (EFS) models using
Bayesian methods via Stan.
}
\section{Main functions}{

\describe{
\item{\code{\link{fit_max_model}}}{Fit EVT or EFS models to data}
\item{\code{\link{get_posterior}}}{Extract posterior samples}
\item{\code{\link{get_max}}}{Estimate maximum size for given return period}
\item{\code{\link{get_pdf}}}{Compute probability density function}
\item{\code{\link{plot_pdf}}}{Visually compare the PDFs of the fitted models}
\item{\code{\link{plot_traceplot}}}{Visualise the fitting process of the models}
}
}

\author{
\strong{Maintainer}: Freddie Heather \email{freddieheather@gmail.com}

}
