% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_evt_model}
\alias{fit_evt_model}
\title{Fit extreme value models}
\usage{
fit_evt_model(
  maxima,
  model_type = c("evt", "efs", "efsmult"),
  chains = 4,
  iter_warmup = 9000,
  iter_sampling = 1000
)
}
\arguments{
\item{maxima}{Vector or list of maximum values}

\item{model_type}{Character: "evt", "efs", or "efsmult"}

\item{chains}{Number of MCMC chains (default: 4)}

\item{iter_warmup}{Number of warmup iterations (default: 9000)}

\item{iter_sampling}{Number of sampling iterations (default: 1000)}
}
\value{
CmdStanMCMC object containing fitted model
}
\description{
Fits either EVT (Extreme Value Theory) or EFS (Extreme Fish Size) models
to marine biological data using Bayesian methods
}
