% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_max.R
\name{g}
\alias{g}
\title{PDF of the maximum of n samples}
\usage{
g(x, n, cdf, pdf)
}
\arguments{
\item{x}{Numeric vector of values at which to evaluate the PDF}

\item{n}{Integer, number of samples}

\item{cdf}{Function returning the CDF F(x)}

\item{pdf}{Function returning the PDF f(x)}
}
\value{
Numeric vector of PDF values for the maximum
}
\description{
PDF of the maximum of n samples
}
\details{
Uses the formula: g(x) = n * F(x)^(n-1) * f(x)
}
\examples{
# Standard normal distribution
g(1.5, n = 10, cdf = pnorm, pdf = dnorm)

# Normal with mean = 5, sd = 2
g(6, n = 10, cdf = \(x) pnorm(x, 5, 2), pdf = \(x) dnorm(x, 5, 2))

# Exponential with rate = 0.5
g(2, n = 5, cdf = \(x) pexp(x, 0.5), pdf = \(x) dexp(x, 0.5))

# Beta distribution
g(0.7, n = 8, cdf = \(x) pbeta(x, 2, 5), pdf = \(x) dbeta(x, 2, 5))
}
