% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{G_max}
\alias{G_max}
\title{CDF of maxima}
\usage{
G_max(x, distr, n, par1, par2)
}
\arguments{
\item{x}{Vector of values}

\item{distr}{Distribution name (e.g., "tnorm")}

\item{n}{Sample size}

\item{par1}{First parameter of underlying distribution}

\item{par2}{Second parameter of underlying distribution}
}
\value{
Vector of cumulative probabilities
}
\description{
Computes the cumulative distribution function of sample maxima
}
