% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{get_cdf}
\alias{get_cdf}
\title{Get cumulative distribution function}
\usage{
get_cdf(fit, xmin = 0, xmax = 300, xstep = 1, ci = 0.8)
}
\arguments{
\item{fit}{Fitted model object}

\item{xmin, xmax}{Range bounds}

\item{xstep}{Step size}

\item{ci}{Credible interval width}
}
\value{
Tibble with size and CDF estimates
}
\description{
Computes CDF across a range of values
}
