% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max.R
\name{get_max}
\alias{get_max}
\title{Estimate maximum size}
\usage{
get_max(fit, ci = 0.8, k = 20)
}
\arguments{
\item{fit}{Named list of CmdStanMCMC objects}

\item{ci}{Credible interval width (default: 0.8 = 80\% credible interval)}

\item{k}{The k-sample LMAX in unit length (e.g., cm) (default: 20)}
}
\value{
Named list of tibbles with maximum length estimates for each model
}
\description{
Estimates maximum size with uncertainty given k-samples
}
