% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdf.R
\name{get_pdf}
\alias{get_pdf}
\title{Get probability density function of the sample maxima}
\usage{
get_pdf(fit, xmin = 0, xmax = 300, xstep = 1, ci = 0.8, k = 20)
}
\arguments{
\item{fit}{Named list of fitted model objects.
Names should be one or more of: "evt", "evt_gumbel", "efs", "efsmm"}

\item{xmin, xmax}{Range bounds}

\item{xstep}{Step size}

\item{ci}{Credible interval width (default = 0.8)}

\item{k}{Integer. Number of samples to be used in the estimation of the EFS PDF (default = 20)}
}
\value{
Named list of tibbles with size and PDF estimates for each model.
}
\description{
Computes PDF across a range of values
}
