% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{get_underlying}
\alias{get_underlying}
\title{Get underlying distribution PDF}
\usage{
get_underlying(fit, xmin = 0, xmax = 300, xstep = 1, ci = 0.8)
}
\arguments{
\item{fit}{Fitted model object}

\item{xmin, xmax}{Range bounds}

\item{xstep}{Step size}

\item{ci}{Credible interval width}
}
\value{
Tibble with size and underlying PDF estimates
}
\description{
Computes the underlying distribution PDF (for EFS models)
}
