% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{inverse_cdf_x}
\alias{inverse_cdf_x}
\title{Quantile function for maxima}
\usage{
inverse_cdf_x(distr, n, par1, par2, p, interval_lwr = 1, interval_upr = 1000)
}
\arguments{
\item{distr}{Distribution name}

\item{n}{Sample size}

\item{par1, par2}{Distribution parameters}

\item{p}{Probability}

\item{interval_lwr, interval_upr}{Search interval bounds}
}
\value{
Quantile value
}
\description{
Computes quantiles for the distribution of sample maxima
}
