% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdf_max.R
\name{invG}
\alias{invG}
\title{Inverse CDF of the maximum of n samples}
\usage{
invG(n, cdf, p = 0.95, lwr = 0, upr = 1000)
}
\arguments{
\item{n}{Integer, number of samples}

\item{cdf}{Function returning the CDF F(x)}

\item{p}{Numeric between 0 and 1, desired quantile, default = 0.95}

\item{lwr}{Integer, lower bound of the derivative calculation, default = 0}

\item{upr}{Integer, upper bound of the derivative calculation, default = 5000}
}
\value{
Numeric vector of CDF values for the maximum
}
\description{
Inverse CDF of the maximum of n samples
}
\details{
Uses the formula: G(x) = F(x)^n
}
\examples{
# Standard normal distribution
invG(n = 1000, cdf = pnorm, p = 0.95)

# Normal with mean = 5, sd = 2
invG(n = 1000, cdf = \(x) pnorm(x, 5, 2), p = 0.95)

# Exponential with rate = 0.5
invG(n = 1000, cdf = \(x) pexp(x, 0.5), p = 0.95)

# Beta distribution
invG(n = 1000, cdf = \(x) pbeta(x, 2, 5), p = 0.95)
}
