% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length.R
\name{kg_to_cm}
\alias{kg_to_cm}
\title{Convert weight to length
Converts weight to length  using the standard allometric relationship weight = a*length^b}
\usage{
kg_to_cm(weight_kg, a = 0.01, b = 3)
}
\arguments{
\item{weight_kg}{Body weight in kilograms}

\item{a}{Allometric coefficient (default: 1)}

\item{b}{Allometric exponent (default: 3)}
}
\value{
Body length in centimetres
}
\description{
Convert weight to length
Converts weight to length  using the standard allometric relationship weight = a*length^b
}
\examples{
kg_to_cm(2.5)
kg_to_cm(c(1.2, 3.4, 0.8), a = 0.01, b = 3)
}
