% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_cdf.R
\name{max_cdf}
\alias{max_cdf}
\title{CDF of the maximum of n samples}
\usage{
max_cdf(l, n, cdf)
}
\arguments{
\item{n}{Single positive integer: number of i.i.d. samples}

\item{cdf}{Function returning the CDF F(x)}

\item{x}{Numeric vector of values at which to evaluate the CDF}
}
\value{
Numeric vector of CDF values for the maximum
}
\description{
#' Computes the cumulative distribution function of the maximum of \code{n} i.i.d. samples
from a distribution with CDF \code{cdf}.
}
\details{
For i.i.d. samples, the CDF of the maximum is
\eqn{G(x) = F(x)^n}, where \eqn{F} is the CDF of the underlying distribution.
}
\examples{
# Standard normal distribution
max_cdf(l = c(1, 1.5), n = 10, cdf = pnorm)

# Normal with mean = 5, sd = 2
max_cdf(l = 6, n = 10, cdf = \(x) pnorm(l, 5, 2))

# Exponential with rate = 0.5
max_cdf(l = 2, n = 5, cdf = \(x) pexp(l, 0.5))

# Beta distribution
max_cdf(l = 0.7, n = 8, cdf = \(l) pbeta(l, 2, 5))
}
