% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_posterior.R
\name{max_posterior}
\alias{max_posterior}
\title{Get the posterior samples of Lmax}
\usage{
max_posterior(fit, ci = 0.8, k = 20, upper_boundary = 500)
}
\arguments{
\item{fit}{Named list of CmdStanMCMC objects}

\item{ci}{Credible interval width (default = 0.8)}

\item{k}{Integer. Number of samples to be used in the estimation of the EFS PDF (default = 20)}

\item{upper_boundary}{Upper search range for the maximum. Default is 500, but increase if expecting larger values.}
}
\value{
Named list of vectors containing posterior samples for each model
}
\description{
Extracts posterior samples from fitted models and estimates the lmax given the parameter posteriors
}
