% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pdf_max}
\alias{pdf_max}
\title{PDF of maxima given distribution parameters}
\usage{
pdf_max(x, distr, n, par1, par2)
}
\arguments{
\item{x}{Vector of values}

\item{distr}{Distribution name (e.g., "tnorm")}

\item{n}{Sample size}

\item{par1}{First parameter of underlying distribution}

\item{par2}{Second parameter of underlying distribution}
}
\value{
Vector of density values
}
\description{
Computes the probability density function of sample maxima
given the underlying distribution
}
