% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_max.R
\name{plot_max}
\alias{plot_max}
\title{Plot model comparison}
\usage{
plot_max(
  fit,
  xmin = 0,
  xmax = 100,
  xstep = 1,
  ci = 0.8,
  k = 20,
  show_title = FALSE,
  show_lines = TRUE,
  col_pallette = c("#2E86AB", "#9e7948ff", "#C77BA0", "#7e348dff"),
  yaxis_title = "Probability density",
  xaxis_title = "Body size, cm",
  xaxis_title_panelB = expression(paste("Estimated ", L[max], ", cm"))
)
}
\arguments{
\item{fit}{Model fit object}

\item{xmin, xmax}{Numeric. Bounds of x-axis (Lmax)}

\item{xstep}{Numeric. Resolution of x-axis, larger step size = lower resolution = faster to run}

\item{ci}{Numeric between 0 and 1. Credible interval width (default = 0.8)}

\item{k}{Integer. The k-sample Lmax for estimation. Note that this is not necessarily the number of sample maxima used to fit the model (default: 20)}

\item{show_title}{Character string. Explanatory title for the vertical lines}

\item{show_lines}{Character string. Vertial lines on plots to indicate Lmax estimates}

\item{col_pallette}{Colour palette for each model used}

\item{yaxis_title}{Title for Y axis (upper plot)}

\item{xaxis_title}{Title for X axis (upper plot)}

\item{xaxis_title_panelB}{Title for X axis (lower plot)}
}
\value{
Combined ggplot object
}
\description{
Creates publication-ready plots comparing EVT and EFS model fits
}
