% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Plot model comparison}
\usage{
plot_model_comparison(
  evt_fit,
  efs_fit,
  data_vector,
  species_name = NULL,
  image_path = NULL,
  ci = 0.8,
  k = 20
)
}
\arguments{
\item{evt_fit}{EVT model fit}

\item{efs_fit}{EFS model fit}

\item{data_vector}{Vector of observed maxima}

\item{species_name}{Scientific name for plot annotation}

\item{image_path}{Optional path to species image}

\item{ci}{Credible interval width (default: 0.8)}

\item{k}{The k-sample LMAX. By default this is set to 20, for ease of comparison between studies and species. Note that this is not the number of sample maxima used to fit the model. (default: 20)}
}
\value{
Combined ggplot object
}
\description{
Creates publication-ready plots comparing EVT and EFS model fits
}
