% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_pdf}
\alias{plot_pdf}
\title{Plot model comparison}
\usage{
plot_pdf(
  fit,
  xmin = 0,
  xmax = 300,
  xstep = 1,
  ci = 0.8,
  k = 20,
  col_pallette = c("#2E86AB", "#C77BA0", "#9e7948ff", "#7e348dff")
)
}
\arguments{
\item{fit}{Model fit object}

\item{xmin, xmax}{Range bounds}

\item{xstep}{Step size}

\item{ci}{Credible interval width (default = 0.8)}

\item{k}{The k-sample LMAX for estimation. Note that this is not mecesarily the number of sample maxima used to fit the model (default: 20)}
}
\value{
Combined ggplot object
}
\description{
Creates publication-ready plots comparing EVT and EFS model fits
}
