% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traceplot.R
\name{plot_traceplot}
\alias{plot_traceplot}
\title{Create traceplot for MCMC diagnostics}
\usage{
plot_traceplot(fit)
}
\arguments{
\item{fit}{Fitted model object}
}
\value{
ggplot object
}
\description{
Creates traceplots for model parameters to assess MCMC convergence
}
