% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cmdstan.R
\name{setup_cmdstan}
\alias{setup_cmdstan}
\title{One-time setup for CmdStan}
\usage{
setup_cmdstan(
  version = NULL,
  cores = max(1L, parallel::detectCores(logical = FALSE))
)
}
\value{
Invisibly returns the CmdStan version.
}
\description{
Checks whether CmdStan is available and installs it if not.
}
