% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{tnorm}
\alias{tnorm}
\alias{dtnorm}
\alias{ptnorm}
\alias{qtnorm}
\alias{rtnorm}
\title{Truncated normal distribution (lower bound at 0)}
\usage{
dtnorm(x, mean, sd)

ptnorm(q, mean = 0, sd = 1)

qtnorm(p, mean, sd)

rtnorm(n, mean = 0, sd = 1)
}
\arguments{
\item{x, q}{Vector of quantiles}

\item{mean}{Vector of means}

\item{sd}{Vector of standard deviations}

\item{p}{Vector of probabilities}

\item{n}{Number of observations}
}
\value{
\code{dtnorm} gives the density, \code{ptnorm} gives the distribution
function, \code{qtnorm} gives the quantile function, and \code{rtnorm}
generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation
for the normal distribution truncated at zero from below.
}
\details{
These are convenience wrappers around \code{\link[truncnorm]{dtruncnorm}}
and related functions, with the lower bound fixed at 0.
}
\examples{
# Density at x = 1
dtnorm(1, mean = 0, sd = 1)

# Generate 100 random values
rtnorm(100, mean = 10, sd = 2)
}
