% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{truncated_normal}
\alias{truncated_normal}
\alias{dtnorm}
\alias{ptnorm}
\alias{qtnorm}
\alias{rtnorm}
\title{Truncated normal distribution functions}
\usage{
dtnorm(x, par1, par2)

ptnorm(x, par1, par2)

qtnorm(x, par1, par2)

rtnorm(x, par1, par2)
}
\arguments{
\item{x, q}{Vector of quantiles}

\item{par1}{Mean parameter}

\item{par2}{Standard deviation parameter}

\item{p}{Vector of probabilities}

\item{n}{Number of observations}
}
\value{
Numeric vector of densities, probabilities, quantiles, or random values
}
\description{
Density, distribution, quantile, and random generation functions
for truncated normal distribution (truncated at 0)
}
