% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{est_max}
\alias{est_max}
\title{Estimate maximum size}
\usage{
est_max(fit, ci = 0.8, k)
}
\arguments{
\item{fit}{Fitted model object}

\item{ci}{Credible interval width (default: 0.8)}

\item{k}{Return period}
}
\value{
Vector of c(estimate, lower CI, upper CI)
}
\description{
Estimates maximum size given return period k
}
