% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_analysis.R
\name{get_posterior}
\alias{get_posterior}
\title{Extract posterior samples}
\usage{
get_posterior(fit)
}
\arguments{
\item{fit}{CmdStanMCMC object from fit_evt_model}
}
\value{
Tibble containing posterior samples
}
\description{
Extracts posterior samples from fitted model as a tibble
}
