% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{gev_vectorised}
\alias{gev_vectorised}
\alias{dgev_v}
\alias{pgev_v}
\alias{qgev_v}
\title{Vectorised GEV distribution functions}
\usage{
dgev_v(x, loc = 0, scale = 1, shape = 0, ...)

pgev_v(q, loc = 0, scale = 1, shape = 0, ...)

qgev_v(p, loc = 0, scale = 1, shape = 0, ...)
}
\arguments{
\item{x, q}{Vector of quantiles}

\item{loc, scale, shape}{Location, scale and shape parameters}

\item{...}{Additional arguments passed to underlying evd functions}
}
\value{
\code{dgev} gives the density and \code{pgev} gives the distribution function.
}
\description{
Vectorised versions of the generalised extreme value distribution functions
from the evd package.
}
\details{
These are vectorised wrappers around \code{\link[evd]{dgev}} and
\code{\link[evd]{pgev}}, allowing all parameters to be vectors.
}
\seealso{
\code{\link[evd]{gev}} for the original non-vectorised functions
}
