% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{gumbel_vectorised}
\alias{gumbel_vectorised}
\alias{dgumbel_v}
\alias{pgumbel_v}
\alias{qgumbel_v}
\title{Vectorised Gumbel distribution functions}
\usage{
dgumbel_v(x, loc = 0, scale = 1, log = FALSE)

pgumbel_v(q, loc = 0, scale = 1, lower.tail = TRUE)

qgumbel_v(p, loc = 0, scale = 1, lower.tail = TRUE)
}
\arguments{
\item{x, q}{Vector of quantiles}

\item{loc, scale}{Location and scale parameters}

\item{...}{Additional arguments passed to underlying evd functions}
}
\value{
\code{dgumbel} gives the density and \code{dgumbel} gives the distribution
function.
}
\description{
Vectorised versions of the Gumbel distribution functions from the evd package.
}
\details{
These are vectorised wrappers around \code{\link[evd]{dgumbel}} and
\code{\link[evd]{pgumbel}}, allowing all parameters to be vectors.
}
\seealso{
\code{\link[evd]{gumbel}} for the original non-vectorised functions
}
