% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{kg_to_cm}
\alias{kg_to_cm}
\title{Convert weight to length for fish}
\usage{
kg_to_cm(weight_kg, a = 0.04478, b = 2.673)
}
\arguments{
\item{weight_kg}{Fish weight in kilograms}

\item{a}{Allometric coefficient (default: 0.04478)}

\item{b}{Allometric exponent (default: 2.673)}
}
\value{
Length in centimetres
}
\description{
Converts fish weight to length using allometric relationships
}
\examples{
kg_to_cm(2.5)
kg_to_cm(c(1.2, 3.4, 0.8), a = 0.01, b = 3)
}
