% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmax_posterior.R
\name{lmax_posterior}
\alias{lmax_posterior}
\title{Get the posterior samples of Lmax}
\usage{
lmax_posterior(fit, ci = 0.8, k = 20)
}
\arguments{
\item{fit}{Named list of CmdStanMCMC objects}
}
\value{
Named list of tibbles containing posterior samples for each model
}
\description{
Extracts posterior samples from fitted models and estimates the lmax given the parameter posteriors
}
