% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Plot model comparison}
\usage{
plot_model_comparison(
  evt_fit,
  efs_fit,
  data_vector,
  species_name = NULL,
  image_path = NULL
)
}
\arguments{
\item{evt_fit}{EVT model fit}

\item{efs_fit}{EFS model fit}

\item{data_vector}{Vector of observed maxima}

\item{species_name}{Scientific name for plot annotation}

\item{image_path}{Optional path to species image}
}
\value{
Combined ggplot object
}
\description{
Creates publication-ready plots comparing EVT and EFS model fits
}
