% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{traceplot}
\alias{traceplot}
\title{Create trace plot for MCMC diagnostics}
\usage{
traceplot(fit)
}
\arguments{
\item{fit}{Fitted model object}
}
\value{
ggplot object
}
\description{
Creates trace plots for model parameters to assess MCMC convergence
}
